+++ src/passwdcalc.js
@@ -5,7 +5,7 @@ class PasswdCalc {
         this.user = config.user;
         this.realm = config.realm;
         this.serialNumber = config.serialNumber;
-    };
+    }
 
     async getPasswd() {
         try {
@@ -13,43 +13,28 @@ class PasswdCalc {
             const inputString = crypto.createHash('md5').update(hashString).digest('hex');
             const digestIterator = this.digestSnippet(inputString);
 
-            const counters = inputString
-                .split("")
-                .reduce((previousValue, currentValue, currentIndex, array) => {
-                    switch (currentValue) {
-                        case "0":
-                            previousValue[0] += 1;
-                            break;
-                        case "1":
-                            previousValue[1] += 1;
-                            break;
-                    };
-                    return previousValue;
-                }, [0, 0]);
+            const counters = inputString.split("").reduce((acc, val) => {
+                if (val === "0") acc[0]++;
+                else if (val === "1") acc[1]++;
+                return acc;
+            }, [0, 0]);
 
-            let countZero = counters[0];
-            let countOne = counters[1];
+            let [countZero, countOne] = counters;
             let password = '';
-            for (const cc of digestIterator) {
-                if (countZero == 3 || countZero == 6 || countZero == 9) {
-                    countZero--;
-                };
-                if (countZero > 20) {
-                    countZero = 20;
-                };
-                if (countZero < 0) {
-                    countZero = 0;
-                };
 
-                if (countOne == 9 || countOne == 15) {
-                    countOne--;
-                };
-                if (countOne > 26) {
-                    countOne = 26;
-                };
-                if (countOne < 0) {
-                    countOne = 0;
-                };
+            const MAX_ZERO = 20;
+            const MAX_ONE = 26;
+            const SKIP_ZERO = [3, 6, 9];
+            const SKIP_ONE = [9, 15];
+
+            for (const cc of digestIterator) {
+                if (SKIP_ZERO.includes(countZero)) countZero--;
+                if (countZero > MAX_ZERO) countZero = MAX_ZERO;
+                if (countZero < 0) countZero = 0;
+
+                if (SKIP_ONE.includes(countOne)) countOne--;
+                if (countOne > MAX_ONE) countOne = MAX_ONE;
+                if (countOne < 0) countOne = 0;
 
                 switch (cc) {
                     case "0":
@@ -63,22 +48,20 @@ class PasswdCalc {
                     default:
                         password += cc;
                         break;
-                };
-            };
+                }
+            }
 
             return password;
         } catch (error) {
-            throw new Error(`Generate password error: ${error}`);
-        };
-    };
+            throw new Error(`Generate password error: ${error.message}`);
+        }
+    }
 
     *digestSnippet(inputString) {
-        let iterationCount = 0;
-        const length = inputString.length;
-        for (let i = length - 1; i > length - 9; i--) {
+        for (let i = inputString.length - 1; i > inputString.length - 9; i--) {
             yield inputString[i];
-        };
-        return iterationCount;
-    };
-};
+        }
+    }
+}
+
 export default PasswdCalc;