+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.14.0] - (18.04.2025)
+
+## Changes
+
+- added multiphase support for q-relay sensor
+- digestauth refactor
+- passwdcsalc refactor
+- envoytoken refactor
+- cleanup
+
 ## [9.13.4] - (18.04.2025)
 
 ## Changes